Clean up in grids

+++ src/grids/linear_grid.cpp
@@ -13,7 +13,7 @@ linear_grid::linear_grid(value_type first, value_type last, size_type size) {
 
 void linear_grid::reset(value_type first, value_type last, size_type size) {
     grid_base::reset(first, last, size);
-    step_ = (last_ - first_) / (size_ - 1);
+    step_ = (last_ - first_) / (static_cast<double>(size_) - 1.0);
 }
 
 } // namespace simplemc
